+++ test/jest/popupElement.test.js
@@ -780,24 +780,24 @@ describe("popupElement", () => {
     );
     await nextFrame();
     expect(el.outerHTML).toMatchInlineSnapshot(
-      `"<wup-popup style="transform: translate(190px, 100px) scaleY(0.055555555555555566); display: block; animation-name: none; transform-origin: bottom;" position="top"><div style="transform: scaleY(17.999999999999996); transform-origin: bottom;"></div><div style="transform: scaleY(17.999999999999996); transform-origin: bottom;"></div></wup-popup>"`
+      `"<wup-popup style="transform: translate(190px, 100px) scaleY(0.055555555555555566); display: block; animation-name: none; transform-origin: bottom;" position="top"><div style="transform: scaleY(17.999999999999996) translateZ(0px); transform-origin: bottom;"></div><div style="transform: scaleY(17.999999999999996) translateZ(0px); transform-origin: bottom;"></div></wup-popup>"`
     );
 
     await nextFrame();
     expect(el.outerHTML).toMatchInlineSnapshot(
-      `"<wup-popup style="transform: translate(190px, 100px) scaleY(0.1111111111111111); display: block; animation-name: none; transform-origin: bottom;" position="top"><div style="transform: scaleY(9); transform-origin: bottom;"></div><div style="transform: scaleY(9); transform-origin: bottom;"></div></wup-popup>"`
+      `"<wup-popup style="transform: translate(190px, 100px) scaleY(0.1111111111111111); display: block; animation-name: none; transform-origin: bottom;" position="top"><div style="transform: scaleY(9) translateZ(0px); transform-origin: bottom;"></div><div style="transform: scaleY(9) translateZ(0px); transform-origin: bottom;"></div></wup-popup>"`
     );
 
     el.$hide();
     expect(el.$isOpen).toBe(true); // because $hide is async
     await Promise.resolve();
     expect(el.outerHTML).toMatchInlineSnapshot(
-      `"<wup-popup style="transform: translate(190px, 100px) scaleY(0.1111111111111111); display: block; animation-name: none; transform-origin: bottom;" position="top" hide=""><div style="transform: scaleY(9); transform-origin: bottom;"></div><div style="transform: scaleY(9); transform-origin: bottom;"></div></wup-popup>"`
+      `"<wup-popup style="transform: translate(190px, 100px) scaleY(0.1111111111111111); display: block; animation-name: none; transform-origin: bottom;" position="top" hide=""><div style="transform: scaleY(9) translateZ(0px); transform-origin: bottom;"></div><div style="transform: scaleY(9) translateZ(0px); transform-origin: bottom;"></div></wup-popup>"`
     );
 
     await nextFrame(2);
     expect(el.outerHTML).toMatchInlineSnapshot(
-      `"<wup-popup style="transform: translate(190px, 100px) scaleY(0.05555555555555553); display: block; animation-name: none; transform-origin: bottom;" position="top" hide=""><div style="transform: scaleY(18.000000000000007); transform-origin: bottom;"></div><div style="transform: scaleY(18.000000000000007); transform-origin: bottom;"></div></wup-popup>"`
+      `"<wup-popup style="transform: translate(190px, 100px) scaleY(0.05555555555555553); display: block; animation-name: none; transform-origin: bottom;" position="top" hide=""><div style="transform: scaleY(18.000000000000007) translateZ(0px); transform-origin: bottom;"></div><div style="transform: scaleY(18.000000000000007) translateZ(0px); transform-origin: bottom;"></div></wup-popup>"`
     );
 
     await nextFrame();
@@ -824,7 +824,7 @@ describe("popupElement", () => {
     );
     await nextFrame();
     expect(el.outerHTML).toMatchInlineSnapshot(
-      `"<wup-popup style="transform: translate(140px, 150px) scaleY(0.055555555555555525); display: block; animation-name: none; transform-origin: top;" position="bottom"><div style="transform: scaleY(18.00000000000001); transform-origin: bottom;"></div><div style="transform: scaleY(18.00000000000001); transform-origin: bottom;"></div></wup-popup>"`
+      `"<wup-popup style="transform: translate(140px, 150px) scaleY(0.055555555555555525); display: block; animation-name: none; transform-origin: top;" position="bottom"><div style="transform: scaleY(18.00000000000001) translateZ(0px); transform-origin: bottom;"></div><div style="transform: scaleY(18.00000000000001) translateZ(0px); transform-origin: bottom;"></div></wup-popup>"`
     );
 
     // checking how it works if options is changed during the animation
@@ -832,30 +832,30 @@ describe("popupElement", () => {
     await h.wait(); // options is async
     await nextFrame();
     expect(el.outerHTML).toMatchInlineSnapshot(
-      `"<wup-popup style="transform: translate(140px, 100px) scaleY(0.055555555555555525); transform-origin: bottom; display: block; animation-name: none;" position="top"><div style="transform: scaleY(18.00000000000001); transform-origin: bottom;"></div><div style="transform: scaleY(18.00000000000001); transform-origin: bottom;"></div></wup-popup>"`
+      `"<wup-popup style="transform: translate(140px, 100px) scaleY(0.055555555555555525); transform-origin: bottom; display: block; animation-name: none;" position="top"><div style="transform: scaleY(18.00000000000001) translateZ(0px); transform-origin: bottom;"></div><div style="transform: scaleY(18.00000000000001) translateZ(0px); transform-origin: bottom;"></div></wup-popup>"`
     );
     await nextFrame();
     expect(el.outerHTML).toMatchInlineSnapshot(
-      `"<wup-popup style="transform: translate(140px, 100px) scaleY(0.11111111111111105); transform-origin: bottom; display: block; animation-name: none;" position="top"><div style="transform: scaleY(9.000000000000005); transform-origin: bottom;"></div><div style="transform: scaleY(9.000000000000005); transform-origin: bottom;"></div></wup-popup>"`
+      `"<wup-popup style="transform: translate(140px, 100px) scaleY(0.11111111111111105); transform-origin: bottom; display: block; animation-name: none;" position="top"><div style="transform: scaleY(9.000000000000005) translateZ(0px); transform-origin: bottom;"></div><div style="transform: scaleY(9.000000000000005) translateZ(0px); transform-origin: bottom;"></div></wup-popup>"`
     );
     el.$hide(); // force to hide during the show-animation
     await nextFrame();
     expect(el.outerHTML).toMatchInlineSnapshot(
-      `"<wup-popup style="transform: translate(140px, 100px) scaleY(0.11111111111111105); transform-origin: bottom; display: block; animation-name: none;" position="top" hide=""><div style="transform: scaleY(9.000000000000005); transform-origin: bottom;"></div><div style="transform: scaleY(9.000000000000005); transform-origin: bottom;"></div></wup-popup>"`
+      `"<wup-popup style="transform: translate(140px, 100px) scaleY(0.11111111111111105); transform-origin: bottom; display: block; animation-name: none;" position="top" hide=""><div style="transform: scaleY(9.000000000000005) translateZ(0px); transform-origin: bottom;"></div><div style="transform: scaleY(9.000000000000005) translateZ(0px); transform-origin: bottom;"></div></wup-popup>"`
     );
 
     await nextFrame();
     expect(el.outerHTML).toMatchInlineSnapshot(
-      `"<wup-popup style="transform: translate(140px, 100px) scaleY(0.05555555555555543); transform-origin: bottom; display: block; animation-name: none;" position="top" hide=""><div style="transform: scaleY(18.000000000000043); transform-origin: bottom;"></div><div style="transform: scaleY(18.000000000000043); transform-origin: bottom;"></div></wup-popup>"`
+      `"<wup-popup style="transform: translate(140px, 100px) scaleY(0.05555555555555543); transform-origin: bottom; display: block; animation-name: none;" position="top" hide=""><div style="transform: scaleY(18.000000000000043) translateZ(0px); transform-origin: bottom;"></div><div style="transform: scaleY(18.000000000000043) translateZ(0px); transform-origin: bottom;"></div></wup-popup>"`
     );
     el.$show(); // force to show during the hide-animation
     await nextFrame();
     expect(el.outerHTML).toMatchInlineSnapshot(
-      `"<wup-popup style="transform: translate(140px, 100px) scaleY(0.05555555555555543); transform-origin: bottom; display: block; animation-name: none;" position="top"><div style="transform: scaleY(18.000000000000043); transform-origin: bottom;"></div><div style="transform: scaleY(18.000000000000043); transform-origin: bottom;"></div></wup-popup>"`
+      `"<wup-popup style="transform: translate(140px, 100px) scaleY(0.05555555555555543); transform-origin: bottom; display: block; animation-name: none;" position="top"><div style="transform: scaleY(18.000000000000043) translateZ(0px); transform-origin: bottom;"></div><div style="transform: scaleY(18.000000000000043) translateZ(0px); transform-origin: bottom;"></div></wup-popup>"`
     );
     await nextFrame();
     expect(el.outerHTML).toMatchInlineSnapshot(
-      `"<wup-popup style="transform: translate(140px, 100px) scaleY(0.11111111111111105); transform-origin: bottom; display: block; animation-name: none;" position="top"><div style="transform: scaleY(9.000000000000005); transform-origin: bottom;"></div><div style="transform: scaleY(9.000000000000005); transform-origin: bottom;"></div></wup-popup>"`
+      `"<wup-popup style="transform: translate(140px, 100px) scaleY(0.11111111111111105); transform-origin: bottom; display: block; animation-name: none;" position="top"><div style="transform: scaleY(9.000000000000005) translateZ(0px); transform-origin: bottom;"></div><div style="transform: scaleY(9.000000000000005) translateZ(0px); transform-origin: bottom;"></div></wup-popup>"`
     );
 
     // checking when element is removed